// src/API.ts
export type CreateTodoInput = {
  id?: string | null,
  name: string,
  status: TodoStatus,
  deadLine?: string | null,
  description?: string | null,
};

export enum TodoStatus {
  NONE = "NONE",
  BACKLOG = "BACKLOG",
  DOING = "DOING",
  REVIEW = "REVIEW",
  DONE = "DONE",
}


export type ModelTodoConditionInput = {
  name?: ModelStringInput | null,
  status?: ModelTodoStatusInput | null,
  deadLine?: ModelStringInput | null,
  description?: ModelStringInput | null,
  and?: Array< ModelTodoConditionInput | null > | null,
  or?: Array< ModelTodoConditionInput | null > | null,
  not?: ModelTodoConditionInput | null,
};

export type ModelStringInput = {
  ne?: string | null,
  eq?: string | null,
  le?: string | null,
  lt?: string | null,
  ge?: string | null,
  gt?: string | null,
  contains?: string | null,
  notContains?: string | null,
  between?: Array< string | null > | null,
  beginsWith?: string | null,
  attributeExists?: boolean | null,
  attributeType?: ModelAttributeTypes | null,
  size?: ModelSizeInput | null,
};
// ...
export type ModelTodoStatusInput = {
  eq?: TodoStatus | null,
  ne?: TodoStatus | null,
};

export type UpdateTodoInput = {
  id: string,
  name?: string | null,
  status?: TodoStatus | null,
  deadLine?: string | null,
  description?: string | null,
};

export type DeleteTodoInput = {
  id?: string | null,
};
// ...
export type CreateTodoMutationVariables = {
  input: CreateTodoInput,
  condition?: ModelTodoConditionInput | null,
};

export type CreateTodoMutation = {
  createTodo:  {
    __typename: "Todo",
    id: string,
    name: string,
    status: TodoStatus,
    deadLine: string | null,
    description: string | null,
    createdAt: string,
    updatedAt: string,
    owner: string | null,
  } | null,
};

export type UpdateTodoMutationVariables = {
  input: UpdateTodoInput,
  condition?: ModelTodoConditionInput | null,
};

export type UpdateTodoMutation = {
  updateTodo:  {
    __typename: "Todo",
    id: string,
    name: string,
    status: TodoStatus,
    deadLine: string | null,
    description: string | null,
    createdAt: string,
    updatedAt: string,
    owner: string | null,
  } | null,
};

export type DeleteTodoMutationVariables = {
  input: DeleteTodoInput,
  condition?: ModelTodoConditionInput | null,
};

export type DeleteTodoMutation = {
  deleteTodo:  {
    __typename: "Todo",
    id: string,
    name: string,
    status: TodoStatus,
    deadLine: string | null,
    description: string | null,
    createdAt: string,
    updatedAt: string,
    owner: string | null,
  } | null,
};